Added mechanism to move in-use extension out the way

+++ src/File/SudoUnlink.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Php\Pie\File;
 
+use Composer\Util\Platform;
 use Php\Pie\Util\CaptureErrors;
 use Php\Pie\Util\Process;
 use Symfony\Component\Process\Exception\ProcessFailedException;
@@ -12,6 +13,7 @@ use function file_exists;
 use function is_writable;
 use function unlink;
 
+/** @internal This is not public API for PIE, so should not be depended upon unless you accept the risk of BC breaks */
 final class SudoUnlink
 {
     public static function singleFile(string $filename): void
@@ -20,7 +22,7 @@ final class SudoUnlink
             return;
         }
 
-        if (is_writable($filename)) {
+        if (! Platform::isWindows() && is_writable($filename)) {
             $capturedErrors   = [];
             $unlinkSuccessful = CaptureErrors::for(
                 static fn () => unlink($filename),
@@ -30,6 +32,12 @@ final class SudoUnlink
             if (! $unlinkSuccessful || file_exists($filename)) {
                 throw FailedToUnlinkFile::fromUnlinkErrors($filename, $capturedErrors);
             }
+
+            return;
+        }
+
+        if (Platform::isWindows()) {
+            WindowsDelete::usingMoveToTemp($filename);
 
             return;
         }
